+++ src/internals/cli/index.js
@@ -1,13 +1,10 @@
 // @flow
 
-import chalk from 'chalk';
 import cli from 'yargs';
 import requireAll from 'require-all';
 import framework from '../../shared/framework-metadata';
 import { getDefault } from '../../shared/util/ModuleUtil';
 import levels from '../../shared/logger/levels';
-
-chalk.enabled = true;
 
 process.on('unhandledRejection', reason => {
   console.error('Unhandled Rejections:');
@@ -47,29 +44,9 @@ for (const file of Object.keys(commands)) {
 cli.demandCommand(1).recommendCommands().argv;
 
 /*
- ## `build`
-
- Build DLLs
-
- `cross-env NODE_ENV=production webpack --config framework/internals/webpack/webpack.prod.babel.js --color -p`
-
- ## `build dll`
-
- `babel-node -- ./framework/internals/scripts/dependencies.js`
-
  ## `extract-intl`
 
  `babel-node -- ./framework/internals/scripts/extract-intl.js`
 
- ## `generate`
-
- `plop --plopfile internals/generators/index.js`
-
- ## `pagespeed`
-
- `node ./internals/scripts/pagespeed.js`
-
- ## `test`
-
- ## `test --coverage`
+ ## serve
  */