+++ src/internals/webpack/build-webpack-client.js
@@ -1,4 +1,4 @@
-import chalk from 'chalk';
+import { chalkUrl } from '../../shared/chalk';
 import logger from '../../shared/logger';
 import clientWebpackConfig from './webpack.client';
 import compileWebpack from './compile-webpack';
@@ -9,6 +9,6 @@ if (process.env.WATCH === 'true') {
   import('./build-webpack-client-watch');
 } else {
   compileWebpack(clientWebpackConfig, false, () => {
-    logger.info(`Client outputted at ${chalk.blue(clientWebpackConfig.output.path)}.`);
+    logger.info(`Client outputted at ${chalkUrl(clientWebpackConfig.output.path)}.`);
   });
 }